% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bittermelon.R
\name{bm_coords}
\alias{bm_coords}
\title{Convert a 'bittermelon' bitmap list into a set of voxel coordinates}
\usage{
bm_coords(bml, sep = NULL)
}
\arguments{
\item{sep}{x separation between characters. If NULL (the default), then 
separation between characters will be determined from the \code{bml} data}

\item{bnl}{\code{bittermelon} list as created with \code{bittermelon::as_bm_list()}}
}
\value{
data.frame of x,y,z coordinates
}
\description{
Convert a 'bittermelon' bitmap list into a set of voxel coordinates
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
font_file <- system.file("fonts/spleen/spleen-8x16.hex.gz", package = "bittermelon")
font      <- bittermelon::read_hex(font_file)
bml       <- bittermelon::as_bm_list("#RStats!", font = font)
coords    <- bm_coords(bml, font)
coords
\dontshow{\}) # examplesIf}
}
