% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{coord_rotate}
\alias{coord_rotate}
\title{Rotate object around a coordinate axis}
\usage{
coord_rotate(coords, theta, axis = "z")
}
\arguments{
\item{coords}{data.frame with 'x', 'y' and 'z' coordinates}

\item{theta}{angle in radians.}

\item{axis}{axis to rotate around. Default: 'z'. Valid values: 'x', 'y', 'z'}
}
\value{
data.frame of transformed coordinates
}
\description{
Rotate object around a coordinate axis
}
\examples{
obj_letter |>
   coord_rotate(pi/2, 'y') |>
   isocubesGrob() |>
   grid::grid.draw()
}
