% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{coord_translate}
\alias{coord_translate}
\title{Translate object}
\usage{
coord_translate(coords, x = 0, y = 0, z = 0)
}
\arguments{
\item{coords}{data.frame with 'x', 'y' and 'z' coordinates}

\item{x, y, z}{amount to translate along each axis. Default: 0}
}
\value{
data.frame of transformed coordinates
}
\description{
Translate object
}
\examples{
gen_sphere() |>
   coord_translate(x = 20, z = 40) |>
   isocubesGrob(size = 2) |>
   grid::grid.draw()
}
