% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{coords_letter}
\alias{coords_letter}
\title{Voxel coordinates for a sphere}
\format{
An object of class \code{data.frame} with 68 rows and 3 columns.
}
\usage{
coords_letter
}
\description{
Voxel coordinates for a sphere
}
\examples{
head(coords_letter)
cubes <- isocubesGrob(coords_letter, size = 5, y = 0)
grid::grid.newpage()
grid::grid.draw(cubes)
}
\seealso{
Other datasets: 
\code{\link{coords_organic}},
\code{\link{coords_sphere}}
}
\concept{datasets}
\keyword{datasets}
