% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{coords_organic}
\alias{coords_organic}
\title{Voxel coordinates for an organic shape}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 14292 rows and 4 columns.
}
\usage{
coords_organic
}
\description{
Voxel coordinates for an organic shape
}
\examples{
head(coords_organic)
cubes <- isocubesGrob(coords_organic, size = 2)
grid::grid.newpage()
grid::grid.draw(cubes)
}
\seealso{
Other datasets: 
\code{\link{coords_letter}},
\code{\link{coords_sphere}}
}
\concept{datasets}
\keyword{datasets}
