% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{coords_sphere}
\alias{coords_sphere}
\title{Voxel coordinates for a sphere}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 17071 rows and 3 columns.
}
\usage{
coords_sphere
}
\description{
Voxel coordinates for a sphere
}
\examples{
head(coords_sphere)
cubes <- isocubesGrob(coords_sphere, size = 2)
grid::grid.newpage()
grid::grid.draw(cubes)
}
\seealso{
Other datasets: 
\code{\link{coords_letter}},
\code{\link{coords_organic}}
}
\concept{datasets}
\keyword{datasets}
