% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\name{gen_prism}
\alias{gen_prism}
\title{Generate a rectangular prism}
\usage{
gen_prism(x = 5, y = x, z = x)
}
\arguments{
\item{x, y, z}{prism dimensions. Default 5x5x5}
}
\value{
data.frame of voxel coordinates
}
\description{
To simplify implementation, only odd side lengths are generated.
}
\examples{
gen_prism(3, 5, 7) |>
   isocubesGrob() |>
   grid::grid.draw()
}
