% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\name{gen_sphere}
\alias{gen_sphere}
\title{Generate voxel coordinates for a sphere}
\usage{
gen_sphere(r = 10)
}
\arguments{
\item{r}{radius}
}
\value{
data.frame of voxel coordinates
}
\description{
Generate voxel coordinates for a sphere
}
\examples{
gen_sphere(1) |>
  isocubesGrob() |>
  grid::grid.draw()
}
