% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isoaxes.R
\name{isoaxesGrob}
\alias{isoaxesGrob}
\title{Create a grob of isocubes}
\usage{
isoaxesGrob(
  size = 5,
  x = NULL,
  y = NULL,
  default.units = "npc",
  default.units.cube = "mm",
  xyplane = "right",
  handedness = "left",
  verbosity = 0,
  ...
)
}
\arguments{
\item{size}{dimensions of cube i.e. the length of the vertical edge of the cube.
Default: 5mm}

\item{x, y}{the origin of the isometric coordinate system in 'snpc' coordinates.
These values should be given as vanilla floating point values.
Be default the origin is the middle bottom of the graphics device 
i.e. \code{(x, y) = (0.5, 0)}}

\item{default.units}{Default unit for (x,y) position is 'npc'}

\item{default.units.cube}{Default unit for size of a cube is 'mm'}

\item{xyplane}{How is the xyplane oriented with respect to the unit isometric
cube?.  "left", "right", "top" (or "flat").
Default: "right"}

\item{handedness}{How is the z-axis positioned with respect to the xy-plane?
I.e. is this a right-handed or left-handed coordinate system?
Default: "left"}

\item{verbosity}{Verbosity level. Default: 0}

\item{...}{other values passed to \code{gpar()} to set the graphical
parameters e.g. \code{lwd} and \code{col} for the linewidth and colour
of the outline stroke for each cube face.}
}
\value{
grid \code{grob} object
}
\description{
Create a grob of isocubes
}
