% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isocubes.R
\name{isocubesGrob}
\alias{isocubesGrob}
\title{Create a grob of isocubes}
\usage{
isocubesGrob(
  coords,
  fill = "grey90",
  fill_left = NULL,
  fill_right = NULL,
  max_y = 20,
  xc = 0.5,
  yc = 1/max_y,
  occlusion_depth = 2L,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{coords}{data.frame of x,y,z coordinates for the cubes (integer coordinates)}

\item{fill}{fill colour}

\item{fill_left, fill_right}{fill colours for left and right faces of
cube.  If set to NULL (the default) then the left and right 
faces will be darkened versions of the given \code{fill} color.}

\item{max_y}{the number of cubes which fit vertically in a window. Default: 20}

\item{xc, yc}{the origin of the isometric coordinate system in 'snpc' coordinates.
Be default the origin is the middle bottom of the graphics device}

\item{occlusion_depth}{How deep should the occlusion checking go in order to 
find cubes which are not visible.  Higher numbers will take more time,
but will remove move cubes which are not visible.
The default value of 2 is a good tradeoff that works well for solid
structures while removing some hidden cubes from sparse structures.
A value of 0 means to not remove any hidden cubes.}

\item{verbose}{Be verbose? Default FALSE.}

\item{...}{other values passed to \code{gpar()} to set the graphical
parameters e.g. \code{lwd} and \code{col} for the linewidth and colour
of the outline stroke for each cube face.}
}
\description{
Create a grob of isocubes
}
