% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{obj_cube}
\alias{obj_cube}
\title{Voxel coordinates for a small cube}
\format{
An object of class \code{data.frame} with 27 rows and 4 columns.
}
\usage{
obj_cube
}
\description{
Voxel coordinates for a small cube
}
\examples{
head(obj_cube)
isocubesGrob(obj_cube, size = 5, y = 0.05) |>
  grid::grid.draw()
}
\seealso{
Other datasets: 
\code{\link{obj_letter}},
\code{\link{obj_organic}},
\code{\link{obj_sphere}},
\code{\link{obj_test}}
}
\concept{datasets}
\keyword{datasets}
