% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{obj_letter}
\alias{obj_letter}
\title{Voxel coordinates for a sphere}
\format{
An object of class \code{data.frame} with 68 rows and 3 columns.
}
\usage{
obj_letter
}
\description{
Voxel coordinates for a sphere
}
\examples{
head(obj_letter)
cubes <- isocubesGrob(obj_letter, size = 5, y = 0.05)
grid::grid.newpage()
grid::grid.draw(cubes)
}
\seealso{
Other datasets: 
\code{\link{obj_organic}},
\code{\link{obj_sphere}}
}
\concept{datasets}
\keyword{datasets}
