% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{obj_organic}
\alias{obj_organic}
\title{Voxel coordinates for an organic shape}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 14292 rows and 4 columns.
}
\usage{
obj_organic
}
\description{
Voxel coordinates for an organic shape
}
\examples{
head(obj_organic)
cubes <- isocubesGrob(obj_organic, size = 2) |>
  grid::grid.draw()
}
\seealso{
Other datasets: 
\code{\link{obj_letter}},
\code{\link{obj_sphere}}
}
\concept{datasets}
\keyword{datasets}
