% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{obj_sphere}
\alias{obj_sphere}
\title{Voxel coordinates for a sphere}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 17071 rows and 3 columns.
}
\usage{
obj_sphere
}
\description{
Voxel coordinates for a sphere
}
\examples{
head(obj_sphere)
cubes <- isocubesGrob(obj_sphere, size = 2) |>
  grid::grid.draw()
}
\seealso{
Other datasets: 
\code{\link{obj_letter}},
\code{\link{obj_organic}}
}
\concept{datasets}
\keyword{datasets}
