% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{obj_test}
\alias{obj_test}
\title{Voxel coordinates for a test object useful for debugging orientation
and visibility checks}
\format{
An object of class \code{data.frame} with 16 rows and 4 columns.
}
\usage{
obj_test
}
\description{
Voxel coordinates for a test object useful for debugging orientation
and visibility checks
}
\examples{
head(obj_test)
isocubesGrob(obj_test, size = 5, y = 0.05) |>
  grid::grid.draw()
}
\seealso{
Other datasets: 
\code{\link{obj_cube}},
\code{\link{obj_letter}},
\code{\link{obj_organic}},
\code{\link{obj_sphere}}
}
\concept{datasets}
\keyword{datasets}
