% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{organic_coords}
\alias{organic_coords}
\title{Voxel coordinates for an organic shape}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 14292 rows and 4 columns.
}
\usage{
organic_coords
}
\description{
Voxel coordinates for an organic shape
}
\examples{
cubes <- isocubesGrob(organic_coords, size = 2)
grid::grid.newpage()
grid::grid.draw(cubes)
}
\keyword{datasets}
