% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{r_coords}
\alias{r_coords}
\title{Voxel coordinates for a sphere}
\format{
An object of class \code{data.frame} with 68 rows and 3 columns.
}
\usage{
r_coords
}
\description{
Voxel coordinates for a sphere
}
\examples{
head(r_coords)
cubes <- isocubesGrob(r_coords, size = 5, y = 0)
grid::grid.newpage()
grid::grid.draw(cubes)
}
\seealso{
Other datasets: 
\code{\link{organic_coords}},
\code{\link{sphere_coords}}
}
\concept{datasets}
\keyword{datasets}
