% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdf.R
\name{sdc_interpolate}
\alias{sdc_interpolate}
\title{Linearly interpolate between two objects}
\usage{
sdc_interpolate(f1, f2, amount)
}
\arguments{
\item{f1, f2}{field functions}

\item{amount}{proportion of final field which comes from second field. 
\code{amount} should be a numeric value in the range [0,1]}
}
\description{
Linearly interpolate between two objects
}
\seealso{
Other sdf_transforms: 
\code{\link{sdc_intersect_smooth}()},
\code{\link{sdc_intersect}()},
\code{\link{sdc_subtract_smooth}()},
\code{\link{sdc_subtract}()},
\code{\link{sdc_union_smooth}()},
\code{\link{sdc_union}()},
\code{\link{sdt_repeat_infinite}()},
\code{\link{sdt_rotatez}()},
\code{\link{sdt_scale}()},
\code{\link{sdt_translate}()}
}
\concept{sdf_transforms}
