% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdf.R
\name{sdc_intersect}
\alias{sdc_intersect}
\title{Intersection of multiple objects}
\usage{
sdc_intersect(f1, f2)
}
\arguments{
\item{f1, f2}{field functions}
}
\description{
Intersection of multiple objects
}
\seealso{
Other sdf_transforms: 
\code{\link{sdc_interpolate}()},
\code{\link{sdc_intersect_smooth}()},
\code{\link{sdc_subtract_smooth}()},
\code{\link{sdc_subtract}()},
\code{\link{sdc_union_smooth}()},
\code{\link{sdc_union}()},
\code{\link{sdt_repeat_infinite}()},
\code{\link{sdt_rotatez}()},
\code{\link{sdt_scale}()},
\code{\link{sdt_translate}()}
}
\concept{sdf_transforms}
