% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdf-objects.R
\name{sdf_box}
\alias{sdf_box}
\title{Create 2x2x2 cube centred at origin}
\usage{
sdf_box()
}
\value{
SDF field function
}
\description{
Create 2x2x2 cube centred at origin
}
\examples{
sdf_box()
}
\seealso{
Other SDF object creation functions: 
\code{\link{sdf_cylinder}()},
\code{\link{sdf_plane}()},
\code{\link{sdf_sphere}()},
\code{\link{sdf_torus}()}
}
\concept{SDF object creation functions}
