% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdf-objects.R
\name{sdf_cylinder}
\alias{sdf_cylinder}
\title{Create infinitely long cylinder at the origin aligned along z axis. r = 1}
\usage{
sdf_cylinder()
}
\value{
SDF field function
}
\description{
Create infinitely long cylinder at the origin aligned along z axis. r = 1
}
\examples{
sdf_cylinder()
}
\seealso{
Other SDF object creation functions: 
\code{\link{sdf_box}()},
\code{\link{sdf_plane}()},
\code{\link{sdf_sphere}()},
\code{\link{sdf_torus}()}
}
\concept{SDF object creation functions}
