% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdf.R
\name{sdf_intersect_smooth}
\alias{sdf_intersect_smooth}
\title{Intersection of multiple objects}
\usage{
sdf_intersect_smooth(f1, f2, k = 0.5)
}
\arguments{
\item{f1, f2}{field functions}

\item{k}{smoothing or interpolationg parameter in range [0, 1]. Default: 0.5}
}
\description{
Intersection of multiple objects
}
\seealso{
Other sdf_combinators: 
\code{\link{sdf_interpolate}()},
\code{\link{sdf_intersect}()},
\code{\link{sdf_subtract_smooth}()},
\code{\link{sdf_subtract}()},
\code{\link{sdf_union_smooth}()},
\code{\link{sdf_union}()}
}
\concept{sdf_combinators}
