% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdf-transforms.R
\name{sdf_onion}
\alias{sdf_onion}
\title{Create a "thin" version of a given object.}
\usage{
sdf_onion(f, thickness)
}
\arguments{
\item{f}{field function}

\item{thickness}{thickness of shell}
}
\value{
SDF field function
}
\description{
Create a "thin" version of a given object.
}
\examples{
sdf_sphere() |>
   sdf_translate(3, 4, 5)
}
\seealso{
Other SDF transformations: 
\code{\link{sdf_repeat_infinite}()},
\code{\link{sdf_rotatez}()},
\code{\link{sdf_round}()},
\code{\link{sdf_scale}()},
\code{\link{sdf_translate}()}
}
\concept{SDF transformations}
