% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdf-objects.R
\name{sdf_plane}
\alias{sdf_plane}
\title{Create semi-infinite volume encompassing all voxels with negative y values}
\usage{
sdf_plane()
}
\value{
SDF field function
}
\description{
This is equivalent to a plane in the XY axis with its normal facing
in the positive y direction
}
\examples{
sdf_plane()
}
\seealso{
Other SDF object creation functions: 
\code{\link{sdf_box}()},
\code{\link{sdf_cylinder}()},
\code{\link{sdf_sphere}()},
\code{\link{sdf_torus}()}
}
\concept{SDF object creation functions}
