% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdf.R
\name{sdf_render}
\alias{sdf_render}
\title{Render a set of objects in an x,y,z volume from -N:N on each axis}
\usage{
sdf_render(scene, N)
}
\arguments{
\item{scene}{sdf object}

\item{N}{extents}
}
\description{
Render a set of objects in an x,y,z volume from -N:N on each axis
}
