% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdf-transforms.R
\name{sdf_rotatez}
\alias{sdf_rotatez}
\alias{sdf_rotatey}
\alias{sdf_rotatex}
\title{Rotate a signed distance field around the z axis}
\usage{
sdf_rotatez(f, theta)

sdf_rotatey(f, theta)

sdf_rotatex(f, theta)
}
\arguments{
\item{f}{field function}

\item{theta}{angle in radians}
}
\value{
SDF field function
}
\description{
Rotate a signed distance field around the z axis
}
\examples{
sdf_sphere() |>
   sdf_translate(3, 4, 5)
}
\seealso{
Other SDF transformations: 
\code{\link{sdf_onion}()},
\code{\link{sdf_repeat_infinite}()},
\code{\link{sdf_round}()},
\code{\link{sdf_scale}()},
\code{\link{sdf_translate}()}
}
\concept{SDF transformations}
