% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdf-transforms.R
\name{sdf_round}
\alias{sdf_round}
\title{Create a rounded version of a given object.}
\usage{
sdf_round(f, r)
}
\arguments{
\item{f}{field function}

\item{r}{radius of rounding}
}
\value{
SDF field function
}
\description{
Create a rounded version of a given object.
}
\examples{
sdf_sphere() |>
   sdf_translate(3, 4, 5)
}
\seealso{
Other SDF transformations: 
\code{\link{sdf_onion}()},
\code{\link{sdf_repeat_infinite}()},
\code{\link{sdf_rotatez}()},
\code{\link{sdf_scale}()},
\code{\link{sdf_translate}()}
}
\concept{SDF transformations}
