% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdf.R
\name{sdf_scale}
\alias{sdf_scale}
\title{Translate an object}
\usage{
sdf_scale(f, xscale = 1, yscale = xscale, zscale = xscale)
}
\arguments{
\item{f}{field function}

\item{xscale, yscale, zscale}{Scale amount. xscale defaults to 1. yscale
and zscale default to the value of xscale}
}
\description{
Translate an object
}
\seealso{
Other sdf_transforms: 
\code{\link{sdf_interpolate}()},
\code{\link{sdf_intersect}()},
\code{\link{sdf_rotatez}()},
\code{\link{sdf_subtract}()},
\code{\link{sdf_translate}()},
\code{\link{sdf_union}()}
}
\concept{sdf_transforms}
