% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdf.R
\name{sdf_set_colour}
\alias{sdf_set_colour}
\title{Set the colour of an object}
\usage{
sdf_set_colour(f, colour)
}
\arguments{
\item{f}{field function}

\item{colour}{colour}
}
\description{
Set the colour of an object
}
