% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdf.R
\name{sdf_sphere}
\alias{sdf_sphere}
\title{Create sphere centred at origin with r = 1}
\usage{
sdf_sphere(colour = "grey90")
}
\arguments{
\item{colour}{object colour}
}
\description{
Create sphere centred at origin with r = 1
}
\seealso{
Other sdf_objects: 
\code{\link{sdf_box}()},
\code{\link{sdf_cyl}()},
\code{\link{sdf_plane}()},
\code{\link{sdf_torus}()}
}
\concept{sdf_objects}
