% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdf-combine.R
\name{sdf_subtract}
\alias{sdf_subtract}
\title{Subtract on object from another}
\usage{
sdf_subtract(f1, f2)
}
\arguments{
\item{f1, f2}{field functions}
}
\value{
SDF field function
}
\description{
Subtract on object from another
}
\examples{
sdf_sphere() |>
   sdf_translate(3, 4, 5)
}
\seealso{
Other SDF Constructive Solid Geometry Operations: 
\code{\link{sdf_interpolate}()},
\code{\link{sdf_intersect}()},
\code{\link{sdf_intersect_smooth}()},
\code{\link{sdf_subtract_smooth}()},
\code{\link{sdf_union}()},
\code{\link{sdf_union_smooth}()}
}
\concept{SDF Constructive Solid Geometry Operations}
