% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdf-transforms.R
\name{sdf_translate}
\alias{sdf_translate}
\title{Translate an object}
\usage{
sdf_translate(f, x = 0, y = 0, z = 0)
}
\arguments{
\item{f}{field function}

\item{x, y, z}{traslation amount. defaults to 0}
}
\value{
SDF field function
}
\description{
Translate an object
}
\examples{
sdf_sphere() |>
   sdf_translate(3, 4, 5)
}
\seealso{
Other SDF transformations: 
\code{\link{sdf_onion}()},
\code{\link{sdf_repeat_infinite}()},
\code{\link{sdf_rotatez}()},
\code{\link{sdf_round}()},
\code{\link{sdf_scale}()}
}
\concept{SDF transformations}
