% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdf.R
\name{sdf_union}
\alias{sdf_union}
\title{Intersection of multiple objects}
\usage{
sdf_union(...)
}
\arguments{
\item{...}{field functions}
}
\description{
Intersection of multiple objects
}
\seealso{
Other sdf_combinators: 
\code{\link{sdf_interpolate}()},
\code{\link{sdf_intersect_smooth}()},
\code{\link{sdf_intersect}()},
\code{\link{sdf_subtract_smooth}()},
\code{\link{sdf_subtract}()},
\code{\link{sdf_union_smooth}()}
}
\concept{sdf_combinators}
