% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdf.R
\name{sdt_repeat_infinite}
\alias{sdt_repeat_infinite}
\alias{sdt_repeat_finite}
\title{Finite/Infinite repetition}
\usage{
sdt_repeat_infinite(f, x, y = x, z = x)

sdt_repeat_finite(f, x, lx, y = x, ly = lx, z = x, lz = lx)
}
\arguments{
\item{f}{field function}

\item{x, y, z}{repetition distance along each axis}

\item{lx, ly, lz}{repeitition times along each axis}
}
\description{
Finite/Infinite repetition
}
\seealso{
Other sdf_transforms: 
\code{\link{sdc_interpolate}()},
\code{\link{sdc_intersect_smooth}()},
\code{\link{sdc_intersect}()},
\code{\link{sdc_subtract_smooth}()},
\code{\link{sdc_subtract}()},
\code{\link{sdc_union_smooth}()},
\code{\link{sdc_union}()},
\code{\link{sdt_rotatez}()},
\code{\link{sdt_scale}()},
\code{\link{sdt_translate}()}
}
\concept{sdf_transforms}
