% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdf.R
\name{sdt_rotatez}
\alias{sdt_rotatez}
\alias{sdt_rotatey}
\alias{sdt_rotatex}
\title{Rotate a signed distance field around the z axis}
\usage{
sdt_rotatez(f, theta)

sdt_rotatey(f, theta)

sdt_rotatex(f, theta)
}
\arguments{
\item{f}{field function}

\item{theta}{angle in radians}
}
\description{
Rotate a signed distance field around the z axis
}
\seealso{
Other sdf_transforms: 
\code{\link{sdc_interpolate}()},
\code{\link{sdc_intersect_smooth}()},
\code{\link{sdc_intersect}()},
\code{\link{sdc_subtract_smooth}()},
\code{\link{sdc_subtract}()},
\code{\link{sdc_union_smooth}()},
\code{\link{sdc_union}()},
\code{\link{sdt_repeat_infinite}()},
\code{\link{sdt_scale}()},
\code{\link{sdt_translate}()}
}
\concept{sdf_transforms}
