% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdf.R
\name{sdt_scale}
\alias{sdt_scale}
\title{Translate an object}
\usage{
sdt_scale(f, xscale = 1, yscale = xscale, zscale = xscale)
}
\arguments{
\item{f}{field function}

\item{xscale, yscale, zscale}{Scale amount. xscale defaults to 1. yscale
and zscale default to the value of xscale}
}
\description{
Translate an object
}
\seealso{
Other sdf_transforms: 
\code{\link{sdc_interpolate}()},
\code{\link{sdc_intersect_smooth}()},
\code{\link{sdc_intersect}()},
\code{\link{sdc_subtract_smooth}()},
\code{\link{sdc_subtract}()},
\code{\link{sdc_union_smooth}()},
\code{\link{sdc_union}()},
\code{\link{sdt_repeat_infinite}()},
\code{\link{sdt_rotatez}()},
\code{\link{sdt_translate}()}
}
\concept{sdf_transforms}
