% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdf.R
\name{sdt_translate}
\alias{sdt_translate}
\title{Translate an object}
\usage{
sdt_translate(f, x = 0, y = 0, z = 0)
}
\arguments{
\item{f}{field function}

\item{x, y, z}{traslation amount. defaults to 0}
}
\description{
Translate an object
}
\seealso{
Other sdf_transforms: 
\code{\link{sdc_interpolate}()},
\code{\link{sdc_intersect_smooth}()},
\code{\link{sdc_intersect}()},
\code{\link{sdc_subtract_smooth}()},
\code{\link{sdc_subtract}()},
\code{\link{sdc_union_smooth}()},
\code{\link{sdc_union}()},
\code{\link{sdt_repeat_infinite}()},
\code{\link{sdt_rotatez}()},
\code{\link{sdt_scale}()}
}
\concept{sdf_transforms}
