% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sphere_coords}
\alias{sphere_coords}
\title{Voxel coordinates for a sphere}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 17071 rows and 3 columns.
}
\usage{
sphere_coords
}
\description{
Voxel coordinates for a sphere
}
\examples{
cubes <- isocubesGrob(sphere_coords, size = 2)
grid::grid.newpage()
grid::grid.draw(cubes)
}
\keyword{datasets}
