% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdf.R
\name{coords_sphere}
\alias{coords_sphere}
\title{Calculate coordinates for isocubes within a sphere}
\usage{
coords_sphere(x, y, z, r)
}
\arguments{
\item{x, y, z}{coordinate of sphere (integers)}

\item{r}{radius of sphere (integer)}
}
\description{
Calculate coordinates for isocubes within a sphere
}
