% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isocubes-package.R
\docType{package}
\name{isocubes-package}
\alias{isocubes}
\alias{isocubes-package}
\title{isocubes: Visualise Voxel Data with Isometric Cubes}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

A voxel is a representation of a value on a regular, three-dimensional grid; it is the 3D equivalent of a 2D pixel. Voxel data can be visualised with this package using fixed viewpoint isometric cubes for each data point. This package also provides sample voxel data and tools for transforming the data.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/coolbutuseless/isocubes}
  \item Report bugs at \url{https://github.com/coolbutuseless/isocubes/issues}
}

}
\author{
\strong{Maintainer}: Mike Cheng \email{mikefc@coolbutuseless.com} [copyright holder]

}
\keyword{internal}
