% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isocubes-package.R
\docType{package}
\name{isocubes-package}
\alias{isocubes}
\alias{isocubes-package}
\title{isocubes: Voxel Rendering with Isometric Cubes}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Visualising three-dimensional voxel data.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/coolbutuseless/isocubes}
  \item Report bugs at \url{https://github.com/coolbutuseless/isocubes/issues}
}

}
\author{
\strong{Maintainer}: Mike Cheng \email{mikefc@coolbutuseless.com} [copyright holder]

}
\keyword{internal}
