% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isocubes.R
\name{isocubesGrob}
\alias{isocubesGrob}
\title{Create a grob of isocubes}
\usage{
isocubesGrob(
  coords,
  fill = "grey90",
  fill_left = NULL,
  fill_right = NULL,
  ysize = 1/20,
  xo = 0.5,
  yo = ysize,
  occlusion_depth = 2L,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{coords}{data.frame of x,y,z coordinates for the cubes (integer coordinates)}

\item{fill}{fill colour for top face of cube.}

\item{fill_left, fill_right}{fill colours for left and right faces of
cube.  If set to NULL (the default) then the left and right 
faces will be darkened versions of the given \code{fill} color.}

\item{ysize}{the size of a cube as a fraction of the window height. Default: 0.05, 
which means that, at most, 20 cubes could stack from top to bottom and be 
visible.}

\item{xo, yo}{the origin of the isometric coordinate system in 'snpc' coordinates.
These values should be given as vanilla floating point values.
Be default the origin is the middle bottom of the graphics device 
i.e. \code{(xo, yo) = (0.5, 0)}}

\item{occlusion_depth}{How deep should the occlusion checking go in order to 
find cubes which are not visible.  Higher numbers will take more time,
but will remove move cubes which are not visible.
The default value of 2 is a good tradeoff that works well for solid
structures while removing some hidden cubes from sparse structures.
A value of 0 means to not remove any hidden cubes.}

\item{verbose}{Be verbose? Default FALSE.}

\item{...}{other values passed to \code{gpar()} to set the graphical
parameters e.g. \code{lwd} and \code{col} for the linewidth and colour
of the outline stroke for each cube face.}
}
\description{
Create a grob of isocubes
}
