% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdf.R
\name{sdf_box}
\alias{sdf_box}
\title{Create 2x2x2 square centred at origin}
\usage{
sdf_box()
}
\description{
Create 2x2x2 square centred at origin
}
\seealso{
Other sdf_objects: 
\code{\link{sdf_cyl}()},
\code{\link{sdf_plane}()},
\code{\link{sdf_sphere}()},
\code{\link{sdf_torus}()}
}
\concept{sdf_objects}
