% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdf.R
\name{sdf_cyl}
\alias{sdf_cyl}
\title{Create an infinitely long unit cylinder at the origin aligned along z axis}
\usage{
sdf_cyl(colour = "hotpink")
}
\arguments{
\item{colour}{object colour}
}
\description{
Create an infinitely long unit cylinder at the origin aligned along z axis
}
