% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdf.R
\name{sdf_cyl}
\alias{sdf_cyl}
\title{Create infinitely long cylinder at the origin aligned along z axis. r = 1}
\usage{
sdf_cyl()
}
\description{
Create infinitely long cylinder at the origin aligned along z axis. r = 1
}
\seealso{
Other sdf_objects: 
\code{\link{sdf_box}()},
\code{\link{sdf_plane}()},
\code{\link{sdf_sphere}()},
\code{\link{sdf_torus}()}
}
\concept{sdf_objects}
