% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdf.R
\name{sdf_interpolate}
\alias{sdf_interpolate}
\title{Linearly interpolate between two objects}
\usage{
sdf_interpolate(f1, f2, amount)
}
\arguments{
\item{f1, f2}{field functions}

\item{amount}{proportion of final field which comes from second field. 
\code{amount} should be a numeric value in the range [0,1]}
}
\description{
Linearly interpolate between two objects
}
\seealso{
Other sdf_transforms: 
\code{\link{sdf_intersect}()},
\code{\link{sdf_rotatez}()},
\code{\link{sdf_scale}()},
\code{\link{sdf_subtract}()},
\code{\link{sdf_translate}()},
\code{\link{sdf_union}()}
}
\concept{sdf_transforms}
