% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdf.R
\name{sdf_intersect}
\alias{sdf_intersect}
\title{Intersection of multiple objects}
\usage{
sdf_intersect(f1, f2)
}
\arguments{
\item{f1, f2}{field functions}
}
\description{
Intersection of multiple objects
}
\seealso{
Other sdf_transforms: 
\code{\link{sdf_interpolate}()},
\code{\link{sdf_repeat_infinite}()},
\code{\link{sdf_rotatez}()},
\code{\link{sdf_scale}()},
\code{\link{sdf_subtract}()},
\code{\link{sdf_translate}()},
\code{\link{sdf_union}()}
}
\concept{sdf_transforms}
