% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdf-combine.R
\name{sdf_intersect_smooth}
\alias{sdf_intersect_smooth}
\title{Intersection of multiple objects}
\usage{
sdf_intersect_smooth(f1, f2, k = 0.5)
}
\arguments{
\item{f1, f2}{field functions}

\item{k}{smoothing or interpolationg parameter in range [0, 1]. Default: 0.5}
}
\value{
SDF field function
}
\description{
Intersection of multiple objects
}
\examples{
sdf_sphere() |>
   sdf_translate(3, 4, 5)
}
\seealso{
Other SDF Constructive Solid Geometry Operations: 
\code{\link{sdf_interpolate}()},
\code{\link{sdf_intersect}()},
\code{\link{sdf_subtract}()},
\code{\link{sdf_subtract_smooth}()},
\code{\link{sdf_union}()},
\code{\link{sdf_union_smooth}()}
}
\concept{SDF Constructive Solid Geometry Operations}
