% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdf.R
\name{sdf_onion}
\alias{sdf_onion}
\title{Create a "thin" version of a given object.}
\usage{
sdf_onion(f, thickness)
}
\arguments{
\item{f}{field function}

\item{thickness}{thickness of shell}
}
\description{
Create a "thin" version of a given object.
}
\seealso{
Other sdf_transforms: 
\code{\link{sdf_repeat_infinite}()},
\code{\link{sdf_rotatez}()},
\code{\link{sdf_round}()},
\code{\link{sdf_scale}()},
\code{\link{sdf_translate}()}
}
\concept{sdf_transforms}
