% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdf.R
\name{sdf_plane}
\alias{sdf_plane}
\title{Create semi-infinite volume encompassing all voxels with negative y values}
\usage{
sdf_plane()
}
\description{
This is equivalent to a plane in the XY axis with its normal facing
in the positive y direction
}
\seealso{
Other sdf_objects: 
\code{\link{sdf_box}()},
\code{\link{sdf_cyl}()},
\code{\link{sdf_sphere}()},
\code{\link{sdf_torus}()}
}
\concept{sdf_objects}
