% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdf.R
\name{sdf_render}
\alias{sdf_render}
\title{Render a set of objects within a voxel volume}
\usage{
sdf_render(scene, N)
}
\arguments{
\item{scene}{sdf object}

\item{N}{Defines the extents of the x,y,z voxel volume 
i.e. from -N:N along each axis}
}
\value{
Return a data.frame of all the integer x,y,z coordinates which 
              are occupied voxels
}
\description{
Render a set of objects within a voxel volume
}
