% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdf.R
\name{sdf_repeat_infinite}
\alias{sdf_repeat_infinite}
\alias{sdf_repeat_finite}
\title{Finite/Infinite repetition}
\usage{
sdf_repeat_infinite(f, x, y = x, z = x)

sdf_repeat_finite(f, x, lx, y = x, ly = lx, z = x, lz = lx)
}
\arguments{
\item{f}{field function}

\item{x, y, z}{repetition distance along each axis}

\item{lx, ly, lz}{repetition times along each axis}
}
\description{
Finite/Infinite repetition
}
\seealso{
Other sdf_transforms: 
\code{\link{sdf_onion}()},
\code{\link{sdf_rotatez}()},
\code{\link{sdf_round}()},
\code{\link{sdf_scale}()},
\code{\link{sdf_translate}()}
}
\concept{sdf_transforms}
