% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdf.R
\name{sdf_round}
\alias{sdf_round}
\title{Create a rounded version of a given object.}
\usage{
sdf_round(f, r)
}
\arguments{
\item{f}{field function}

\item{r}{radius of rounding}
}
\description{
Create a rounded version of a given object.
}
\seealso{
Other sdf_transforms: 
\code{\link{sdf_onion}()},
\code{\link{sdf_repeat_infinite}()},
\code{\link{sdf_rotatez}()},
\code{\link{sdf_scale}()},
\code{\link{sdf_translate}()}
}
\concept{sdf_transforms}
