% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdf.R
\name{sdf_scale}
\alias{sdf_scale}
\title{Scale an object}
\usage{
sdf_scale(f, xscale = 1, yscale = xscale, zscale = xscale)
}
\arguments{
\item{f}{field function}

\item{xscale, yscale, zscale}{Scale amount. xscale defaults to 1. yscale
and zscale default to the value of xscale}
}
\description{
Scale an object
}
\seealso{
Other sdf_transforms: 
\code{\link{sdf_onion}()},
\code{\link{sdf_repeat_infinite}()},
\code{\link{sdf_rotatez}()},
\code{\link{sdf_round}()},
\code{\link{sdf_translate}()}
}
\concept{sdf_transforms}
