% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdf-objects.R
\name{sdf_sphere}
\alias{sdf_sphere}
\title{Create sphere centered at the origin with r = 1}
\usage{
sdf_sphere()
}
\value{
SDF field function
}
\description{
Create sphere centered at the origin with r = 1
}
\examples{
sdf_sphere()
}
\seealso{
Other SDF object creation functions: 
\code{\link{sdf_box}()},
\code{\link{sdf_cylinder}()},
\code{\link{sdf_plane}()},
\code{\link{sdf_torus}()}
}
\concept{SDF object creation functions}
