% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdf.R
\name{sdf_subtract}
\alias{sdf_subtract}
\title{Subtract on object from another}
\usage{
sdf_subtract(f1, f2, colour_opt = 1L)
}
\arguments{
\item{f1, f2}{field functions}

\item{colour_opt}{how to handle colours
\describe{
\item{0}{Choose the best colour per voxel based upon interplation, blending
     or whatever the operation might be}
\item{1:n}{Use the colour from the nth object only. Default: 1 }
}}
}
\description{
Subtract on object from another
}
\seealso{
Other sdf_combinators: 
\code{\link{sdf_interpolate}()},
\code{\link{sdf_intersect_smooth}()},
\code{\link{sdf_intersect}()},
\code{\link{sdf_subtract_smooth}()},
\code{\link{sdf_union_smooth}()},
\code{\link{sdf_union}()}
}
\concept{sdf_combinators}
