% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdf.R
\name{sdf_subtract}
\alias{sdf_subtract}
\title{Subtract on object from another}
\usage{
sdf_subtract(f1, f2)
}
\arguments{
\item{f1, f2}{field functions}
}
\description{
Subtract on object from another
}
\seealso{
Other sdf_combinators: 
\code{\link{sdf_interpolate}()},
\code{\link{sdf_intersect_smooth}()},
\code{\link{sdf_intersect}()},
\code{\link{sdf_subtract_smooth}()},
\code{\link{sdf_union_smooth}()},
\code{\link{sdf_union}()}
}
\concept{sdf_combinators}
