% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdf.R
\name{sdf_subtract}
\alias{sdf_subtract}
\title{Subtract on object from another}
\usage{
sdf_subtract(f1, f2)
}
\arguments{
\item{f1, f2}{field functions}
}
\description{
Subtract on object from another
}
\seealso{
Other sdf_transforms: 
\code{\link{sdf_interpolate}()},
\code{\link{sdf_intersect}()},
\code{\link{sdf_repeat_infinite}()},
\code{\link{sdf_rotatez}()},
\code{\link{sdf_scale}()},
\code{\link{sdf_translate}()},
\code{\link{sdf_union}()}
}
\concept{sdf_transforms}
