% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdf-objects.R
\name{sdf_torus}
\alias{sdf_torus}
\title{Create torus (donut) at origin}
\usage{
sdf_torus(r1, r2)
}
\arguments{
\item{r1}{radius from centre of torus}

\item{r2}{radius of the ring.    r2 < r1}
}
\value{
SDF field function
}
\description{
Create torus (donut) at origin
}
\examples{
sdf_torus(5, 1)
}
\seealso{
Other SDF object creation functions: 
\code{\link{sdf_box}()},
\code{\link{sdf_cylinder}()},
\code{\link{sdf_plane}()},
\code{\link{sdf_sphere}()}
}
\concept{SDF object creation functions}
