% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdf.R
\name{sdf_torus}
\alias{sdf_torus}
\alias{sdf_sphere}
\alias{sdf_plane}
\alias{sdf_box}
\title{Create a unit shape at the origin}
\usage{
sdf_torus(r1, r2, colour = "hotpink")

sdf_sphere(colour = "hotpink")

sdf_plane(colour = "hotpink")

sdf_box(colour = "hotpink")
}
\arguments{
\item{r1}{radius from centre of torus}

\item{r2}{radius of the ring.    r2 < r1}

\item{colour}{object colour}
}
\description{
Many SDF functions from \url{https://iquilezles.org/articles/distfunctions/}
}
\details{
\describe{
 \item{cylinder}{infinitely long unit cylinder at the origin aligned along z axis}
}
}
