% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdf.R
\name{sdf_torus}
\alias{sdf_torus}
\title{Create torus (donut) at origin}
\usage{
sdf_torus(r1, r2)
}
\arguments{
\item{r1}{radius from centre of torus}

\item{r2}{radius of the ring.    r2 < r1}
}
\description{
Create torus (donut) at origin
}
\seealso{
Other sdf_objects: 
\code{\link{sdf_box}()},
\code{\link{sdf_cyl}()},
\code{\link{sdf_plane}()},
\code{\link{sdf_sphere}()}
}
\concept{sdf_objects}
