% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdf-combine.R
\name{sdf_union}
\alias{sdf_union}
\title{Intersection of multiple objects}
\usage{
sdf_union(...)
}
\arguments{
\item{...}{field functions}
}
\value{
SDF field function
}
\description{
Intersection of multiple objects
}
\examples{
sdf_sphere() |>
   sdf_translate(3, 4, 5)
}
\seealso{
Other SDF Constructive Solid Geometry Operations: 
\code{\link{sdf_interpolate}()},
\code{\link{sdf_intersect}()},
\code{\link{sdf_intersect_smooth}()},
\code{\link{sdf_subtract}()},
\code{\link{sdf_subtract_smooth}()},
\code{\link{sdf_union_smooth}()}
}
\concept{SDF Constructive Solid Geometry Operations}
