% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdf.R
\name{sdf_union_smooth}
\alias{sdf_union_smooth}
\title{Smoothed union of two objects}
\usage{
sdf_union_smooth(f1, f2, k)
}
\arguments{
\item{f1, f2}{field functions}

\item{k}{smoothing parameter}
}
\description{
Smoothed union of two objects
}
\seealso{
Other sdf_transforms: 
\code{\link{sdf_interpolate}()},
\code{\link{sdf_intersect_smooth}()},
\code{\link{sdf_intersect}()},
\code{\link{sdf_onion}()},
\code{\link{sdf_repeat_infinite}()},
\code{\link{sdf_rotatez}()},
\code{\link{sdf_round}()},
\code{\link{sdf_scale}()},
\code{\link{sdf_subtract_smooth}()},
\code{\link{sdf_subtract}()},
\code{\link{sdf_translate}()},
\code{\link{sdf_union}()}
}
\concept{sdf_transforms}
